ProductBuilder consistent with other builders
